Move MainWindow export orchestration into MainWindow.xaml.cpp
