+++ Win/llvc/llvc.vcxproj.filters
@@ -12,7 +12,6 @@
   <ItemGroup>
     <ClCompile Include="pch.cpp" />
     <ClCompile Include="Generated Files\module.g.cpp" />
-    <ClCompile Include="MainWindow.Export.cpp" />
     <ClCompile Include="Project.ixx" />
     <ClCompile Include="Timeline.ixx" />
     <ClCompile Include="Utils.ixx" />
